change value to float

+++ services/rating/internal/pkg/router/payloads.go
@@ -1,7 +1,7 @@
 package router
 
 type createPayload struct {
-	UserID  string `json:"userId"`
-	MovieID string `json:"movieId"`
-	Value   int    `json:"value"`
+	UserID  string  `json:"userId"`
+	MovieID string  `json:"movieId"`
+	Value   float32 `json:"value"`
 }
